Implement enrollment model (Matricula) and its serializer, viewset, and routes.
The Matricula model establishes the relationship between student and
course (Aluno and Curso), defined the latter two as foreign keys in the
database.
The definition of foreign keys is the main difference between to the
implementation of Aluno and Curso.

As usual:

- The model has been defined in `escola/models.py`
- The model table has been defined for Django Admin in `escola/admin.py`
- The serializer  has been defined in `escola/serializers.py`
- The viewset has been defined in `escola/views.py`
- The routes have been defined in `setup/urls.py`

Since this commit introduces a new model (Matricula), the database
migrations have been created as well.
Once applied, the enrollment list should be accessible via
`http://127.0.0.1:8000/matriculas/`.